Move QuestionType to domain layer
